requirements.txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.30.0
chromadb
beautifulsoup4
ctransformers==0.2.5 
sentence-transformers==2.2.2
html2text==2020.1.16
firecrawl-py
openai
pandas
openpyxl
selenium
qdrant-client
torch
fastapi==0.109.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
numpy